[EST-34] Chore: 폴더 분리
